5.20 zmj更新 计算精度更新 个人中心更新

+++ src/main/java/com/zmj/wkt/controller/TbkController.java
@@ -247,9 +247,9 @@ public class TbkController extends CommonController {
         bs_orderform.setSubID("Tbk_"+ UUID.randomUUID().toString().toUpperCase());
         bs_orderform.setState(SysCode.STATE_TO_BE_SENT.getCode());
         bs_orderform.setIsAble(SysCode.IS_ABLE_YES.getCode());
-        bs_orderform.setGPrice(bs_goods.getGTbkPrice().doubleValue());
+        bs_orderform.setGPrice(bs_goods.getGTbkPrice());
         bs_orderform.setSpCount(num_iids.length);
-        bs_orderform.setSpPrice(bs_goods.getGTbkPrice().multiply(BigDecimal.valueOf(num_iids.length)).doubleValue());
+        bs_orderform.setSpPrice(bs_goods.getGTbkPrice().multiply(BigDecimal.valueOf(num_iids.length)));
         bs_orderform.setItemTitle("淘客商品订单");
         bs_orderform.setItemDescription(Arrays.toString(num_iids));
         //申请时间
